Extract os-setup/dependency management into reusable osdep package
Move the OpenStudio + OpenStudio-HPXML dependency manager out of
src/h2k_hpxml/utils/dependencies/ into the standalone, reusable
`openstudio-deps` package (import name `osdep`), consumed here as a path
dependency.

- utils/dependencies/__init__.py is now a backward-compat shim re-exporting
  osdep's public API, so existing imports keep working unchanged.
- _h2k_versions.py loads h2k's pinned versions from
  resources/dependency_versions.json and injects them globally via
  osdep.set_default_config(), so zero-arg helper calls (config/manager.py,
  api.py) resolve to h2k's versions. Removed the now-unused
  [tool.h2k-hpxml.dependencies] pyproject section.
- New cli/os_setup.py wrapper keeps the h2k-specific --setup and
  --test-installation/--test-comprehensive behaviors; os-setup entry point
  repointed to it. Dropped the broken --update-config/--add-to-path flags.
- Fixed pre-existing latent bug: convert.py/resilience.py called undefined
  DependencyManager._get_openstudio_paths(); now use the public helper.
- Generic dependency tests moved to the osdep repo; added
  test_dependencies_shim.py to lock in the shim/injection contract.
