--- src/h2k_hpxml/utils/dependencies/installers/__init__.py
@@ -1,5 +1,0 @@
-"""Installer modules for OpenStudio and HPXML dependencies."""
-
-from .base import BaseInstaller
-
-__all__ = ["BaseInstaller"]
