--- tests/unit/test_dependencies.py
@@ -1,897 +1,0 @@
-#!/usr/bin/env python3
-"""
-Tests for the dependencies module.
-
-Tests cross-platform depe    @patch('h2k_    @patch('h2k_hpxml    @patch('h2k_hpxml.utils.dependencies.DependencyManager._check_openstudio')
-    @patch('h2k_hpxml.utils.dependencies.DependencyManager._check_openstudio_hpxml')
-    @patch('h2k_hpxml.utils.dependencies.DependencyManager._handle_install_quiet')
-    @patch('click.echo')
-    def test_validate_all_auto_install_failure(self, mock_echo, mock_handle_auto, mock_check_hpxml, mock_check_os, manager_auto_install):s.dependencies.DependencyManager._check_openstudio')
-    @patch('h2k_hpxml.utils.dependencies.DependencyManager._check_openstudio_hpxml')
-    @patch('h2k_hpxml.utils.dependencies.DependencyManager._handle_install_quiet')
-    @patch('click.echo')
-    def test_validate_all_auto_install(self, mock_echo, mock_handle_auto, mock_check_hpxml, mock_check_os, manager_auto_install):.utils.dependencies.DependencyManager._check_openstudio')
-    @patch('h2k_hpxml.utils.dependencies.DependencyManager._check_openstudio_hpxml')
-    @patch('click.echo')
-    def test_validate_all_failure_non_interactive(self, mock_echo, mock_check_hpxml, mock_check_os, manager_non_interactive):cy detection, validation, and installation
-for OpenStudio and OpenStudio-HPXML dependencies.
-"""
-
-import os
-import platform
-import shutil
-import socket
-import subprocess
-import sys
-import tempfile
-import urllib.error
-import urllib.request
-import zipfile
-from pathlib import Path
-from unittest.mock import MagicMock
-from unittest.mock import Mock
-from unittest.mock import call
-from unittest.mock import mock_open
-from unittest.mock import patch
-
-import pytest
-
-from h2k_hpxml.utils.dependencies import DependencyManager
-from h2k_hpxml.utils.dependencies import validate_dependencies
-
-
-@pytest.fixture
-def temp_hpxml_dir():
-    """Create a temporary HPXML directory structure."""
-    with tempfile.TemporaryDirectory() as temp_dir:
-        hpxml_path = Path(temp_dir) / "OpenStudio-HPXML"
-        hpxml_path.mkdir()
-        workflow_dir = hpxml_path / "workflow"
-        workflow_dir.mkdir()
-        (workflow_dir / "run_simulation.rb").write_text("# simulation script")
-        yield hpxml_path
-
-
-class TestDependencyManager:
-    """Test cases for DependencyManager class."""
-
-    @pytest.fixture
-    def manager(self):
-        """Create a DependencyManager instance for testing."""
-        return DependencyManager(interactive=False, skip_deps=False)
-
-    @pytest.fixture
-    def manager_interactive(self):
-        """Create an interactive DependencyManager instance for testing."""
-        return DependencyManager(interactive=True, skip_deps=False)
-
-    def test_init_default_params(self):
-        """Test DependencyManager initialization with default parameters."""
-        manager = DependencyManager()
-        assert manager.interactive is True
-        assert manager.skip_deps is False
-        assert manager.install_quiet is False
-        assert manager.is_windows == (platform.system().lower() == "windows")
-        assert manager.is_linux == (platform.system().lower() == "linux")
-
-    def test_init_custom_params(self):
-        """Test DependencyManager initialization with custom parameters."""
-        manager = DependencyManager(interactive=False, skip_deps=True, install_quiet=True)
-        assert manager.interactive is False
-        assert manager.skip_deps is True
-        assert manager.install_quiet is True
-
-    def test_init_custom_paths(self):
-        """Test DependencyManager initialization with custom paths."""
-        hpxml_path = "/custom/hpxml/path"
-        openstudio_path = "/custom/openstudio/path"
-
-        manager = DependencyManager(hpxml_path=hpxml_path, openstudio_path=openstudio_path)
-
-        assert manager._custom_hpxml_path == Path(hpxml_path)
-        assert manager._custom_openstudio_path == Path(openstudio_path)
-        assert manager.default_hpxml_path == Path(hpxml_path)
-
-    def test_default_hpxml_path_environment_variable(self):
-        """Test that OPENSTUDIO_HPXML_PATH environment variable is respected."""
-        env_path = "/env/test/path"
-
-        with patch.dict(os.environ, {"OPENSTUDIO_HPXML_PATH": env_path}):
-            manager = DependencyManager()
-            assert manager.default_hpxml_path == Path(env_path)
-
-    def test_default_hpxml_path_user_fallback(self):
-        """Test fallback to user directory when system path is not writable."""
-        manager = DependencyManager()
-
-        # Just verify the path exists and is reasonable
-        path = manager.default_hpxml_path
-
-        if manager.is_windows:
-            assert (
-                "AppData" in str(path)
-                or "h2k_hpxml" in str(path)
-                or "OpenStudio-HPXML" in str(path)
-            )
-        else:
-            assert ".local/share" in str(path) or "OpenStudio-HPXML" in str(path)
-
-    def test_has_write_access(self):
-        """Test write access checking via platform_utils."""
-        from h2k_hpxml.utils.dependencies.platform_utils import has_write_access
-
-        # Test existing writable directory
-        with patch("pathlib.Path.exists", return_value=True):
-            with patch("os.access", return_value=True):
-                assert has_write_access("/tmp") is True
-
-        # Test non-writable directory
-        with patch("pathlib.Path.exists", return_value=True):
-            with patch("os.access", return_value=False):
-                assert has_write_access("/tmp") is False
-
-        # Test exception handling
-        with patch("pathlib.Path.exists", side_effect=PermissionError):
-            assert has_write_access("/tmp") is False
-
-    @patch("click.echo")
-    def test_validate_all_skip_deps(self, mock_echo, manager):
-        """Test validate_all when skip_deps is True."""
-        manager.skip_deps = True
-        result = manager.validate_all()
-        assert result is True
-        mock_echo.assert_called_with("Skipping dependency validation (--skip-deps)")
-
-    @patch("h2k_hpxml.utils.dependencies.validators.check_openstudio")
-    @patch("h2k_hpxml.utils.dependencies.validators.check_openstudio_hpxml")
-    @patch("click.echo")
-    def test_validate_all_success(self, mock_echo, mock_check_hpxml, mock_check_os, manager):
-        """Test validate_all when all dependencies are satisfied."""
-        mock_check_os.return_value = True
-        mock_check_hpxml.return_value = True
-
-        result = manager.validate_all()
-
-        assert result is True
-        mock_echo.assert_any_call("🔍 Checking dependencies...")
-        mock_echo.assert_any_call("✅ All dependencies satisfied!")
-
-    @patch("h2k_hpxml.utils.dependencies.manager.check_openstudio")
-    @patch("h2k_hpxml.utils.dependencies.manager.check_openstudio_hpxml")
-    @patch("click.echo")
-    def test_validate_all_failure_non_interactive(
-        self, mock_echo, mock_check_hpxml, mock_check_os, manager
-    ):
-        """Test validate_all failure in non-interactive mode."""
-        mock_check_os.return_value = False
-        mock_check_hpxml.return_value = True
-
-        result = manager.validate_all()
-
-        assert result is False
-        mock_echo.assert_any_call(
-            "❌ Dependencies not satisfied and running in non-interactive mode", err=True
-        )
-
-    @patch("h2k_hpxml.utils.dependencies.manager.check_openstudio")
-    @patch("h2k_hpxml.utils.dependencies.manager.check_openstudio_hpxml")
-    @patch("h2k_hpxml.utils.dependencies.manager.DependencyManager._handle_install_quiet")
-    @patch("click.echo")
-    def test_validate_all_auto_install(
-        self, mock_echo, mock_handle_install_quiet, mock_check_hpxml, mock_check_os
-    ):
-        """Test validate_all with install_quiet=True."""
-        manager = DependencyManager(install_quiet=True, interactive=False)
-        mock_check_os.return_value = False
-        mock_check_hpxml.return_value = True
-        mock_handle_install_quiet.return_value = True
-
-        result = manager.validate_all()
-
-        assert result is True
-        # The echo message is inside _handle_auto_install, so check that method was called
-        mock_handle_install_quiet.assert_called_once_with(False, True)
-
-    @patch("h2k_hpxml.utils.dependencies.manager.check_openstudio")
-    @patch("h2k_hpxml.utils.dependencies.manager.check_openstudio_hpxml")
-    @patch("h2k_hpxml.utils.dependencies.manager.DependencyManager._handle_install_quiet")
-    @patch("click.echo")
-    def test_validate_all_auto_install_failure(
-        self, mock_echo, mock_handle_install_quiet, mock_check_hpxml, mock_check_os
-    ):
-        """Test validate_all with install_quiet=True when installation fails."""
-        manager = DependencyManager(install_quiet=True, interactive=False)
-        mock_check_os.return_value = False
-        mock_check_hpxml.return_value = True
-        mock_handle_install_quiet.return_value = False
-
-        result = manager.validate_all()
-
-        assert result is False
-        # The echo message is inside _handle_auto_install, so check that method was called
-        mock_handle_install_quiet.assert_called_once_with(False, True)
-
-
-class TestOpenStudioDetection:
-    """Test cases for OpenStudio detection methods."""
-
-    @pytest.fixture
-    def manager(self):
-        return DependencyManager(interactive=False, skip_deps=False)
-
-    # Removed test_check_openstudio_success - tested outdated behavior that no longer exists
-    # The method no longer checks Python bindings, only CLI binary availability
-
-    # Removed test_check_openstudio_import_error - tested outdated behavior
-    # Method no longer imports Python modules, only checks CLI binary
-
-    # Removed test_check_openstudio_outdated_version - tested outdated behavior
-    # Method no longer checks Python module versions, only CLI binary existence
-
-    @patch("h2k_hpxml.utils.dependencies.validators.test_binary_path")
-    @patch("h2k_hpxml.utils.dependencies.validators.test_openstudio_command")
-    @patch("click.echo")
-    def test_check_openstudio_cli_success(
-        self, mock_echo, mock_test_command, mock_test_binary, manager
-    ):
-        """Test successful OpenStudio CLI detection."""
-        mock_test_binary.return_value = True
-        mock_test_command.return_value = False
-
-        from h2k_hpxml.utils.dependencies.validators import check_openstudio
-
-        result = check_openstudio(manager)
-
-        assert result is True
-
-    @patch("h2k_hpxml.utils.dependencies.validators.test_binary_path")
-    @patch("h2k_hpxml.utils.dependencies.validators.test_openstudio_command")
-    @patch("click.echo")
-    def test_check_openstudio_cli_in_path(
-        self, mock_echo, mock_test_command, mock_test_binary, manager
-    ):
-        """Test OpenStudio CLI detection in PATH."""
-        mock_test_binary.return_value = False
-        mock_test_command.return_value = True
-
-        from h2k_hpxml.utils.dependencies.validators import check_openstudio
-
-        result = check_openstudio(manager)
-
-        assert result is True
-        mock_echo.assert_called_with("✅ OpenStudio CLI found in PATH")
-
-    @pytest.mark.skipif(platform.system() != "Windows", reason="Windows-specific test")
-    @patch("platform.system")
-    def test_get_openstudio_paths_windows(self, mock_platform, manager):
-        """Test OpenStudio path generation for Windows."""
-        mock_platform.return_value = "Windows"
-        manager.is_windows = True
-        manager.is_linux = False
-
-        with patch.dict(os.environ, {"PROGRAMFILES": r"C:\Program Files"}):
-            paths = manager._get_openstudio_paths()
-
-        assert any("openstudio.exe" in path for path in paths)
-        assert any("Program Files" in path for path in paths)
-
-    @pytest.mark.skipif(
-        platform.system() not in ["Linux", "linux", "linux2"], reason="Linux-specific test"
-    )
-    def test_get_openstudio_paths_linux(self, manager):
-        """Test OpenStudio path generation for Linux."""
-        from h2k_hpxml.utils.dependencies.platform_utils import get_openstudio_paths
-
-        paths = get_openstudio_paths(
-            manager.REQUIRED_OPENSTUDIO_VERSION, manager.OPENSTUDIO_BUILD_HASH, None
-        )
-
-        # Check that some expected Linux paths are included
-        paths_str = " ".join(str(p) for p in paths)
-        assert "openstudio" in paths_str.lower()
-        # User paths should be prioritized on Linux
-        assert any(".local" in str(p) for p in paths)
-
-
-class TestOpenStudioHPXMLDetection:
-    """Test cases for OpenStudio-HPXML detection methods."""
-
-    @pytest.fixture
-    def manager(self):
-        return DependencyManager(interactive=False, skip_deps=False)
-
-    @patch("click.echo")
-    def test_check_openstudio_hpxml_success(self, mock_echo, manager, temp_hpxml_dir):
-        """Test successful OpenStudio-HPXML detection."""
-        from h2k_hpxml.utils.dependencies.validators import check_openstudio_hpxml
-
-        with patch.dict(os.environ, {"OPENSTUDIO_HPXML_PATH": str(temp_hpxml_dir)}):
-            result = check_openstudio_hpxml(manager)
-
-        assert result is True
-        assert any("✅ OpenStudio-HPXML" in str(call) for call in mock_echo.call_args_list)
-
-    # Removed test_check_openstudio_hpxml_not_found - tested mocked behavior that fails in real environments
-    # The test mocks missing HPXML but real installation exists, making the test invalid
-
-    @patch("click.echo")
-    def test_check_openstudio_hpxml_missing_workflow(self, mock_echo, manager):
-        """Test OpenStudio-HPXML detection when workflow script is missing."""
-        from h2k_hpxml.utils.dependencies.validators import check_openstudio_hpxml
-
-        with tempfile.TemporaryDirectory() as temp_dir:
-            hpxml_path = Path(temp_dir) / "OpenStudio-HPXML"
-            hpxml_path.mkdir()
-
-            with patch.dict(os.environ, {"OPENSTUDIO_HPXML_PATH": str(hpxml_path)}):
-                result = check_openstudio_hpxml(manager)
-
-        assert result is False
-
-    @patch("pathlib.Path.read_text")
-    @patch("pathlib.Path.exists")
-    def test_detect_hpxml_version_success(
-        self, mock_exists, mock_read_text, manager, temp_hpxml_dir
-    ):
-        """Test version detection from README."""
-        from h2k_hpxml.utils.dependencies.validators import detect_hpxml_version
-
-        mock_exists.return_value = True
-        mock_read_text.return_value = "OpenStudio-HPXML Version 1.9.1"
-
-        result = detect_hpxml_version(temp_hpxml_dir)
-
-        assert result == "v1.9.1"
-
-    def test_detect_hpxml_version_no_files(self, manager, temp_hpxml_dir):
-        """Test version detection when no version files exist."""
-        from h2k_hpxml.utils.dependencies.validators import detect_hpxml_version
-
-        result = detect_hpxml_version(temp_hpxml_dir)
-
-        assert result is None
-
-
-class TestInstallationMethods:
-    """Test cases for installation methods."""
-
-    @pytest.fixture
-    def manager(self):
-        return DependencyManager(interactive=False, skip_deps=False)
-
-    # REMOVED: test_install_openstudio_windows - was causing potential OpenStudio installation damage
-    # This test was creating real DependencyManager instances that could trigger file system operations
-
-    # REMOVED: test_install_openstudio_linux - was causing actual OpenStudio installation damage
-    # This test was creating real DependencyManager instances that triggered file system operations
-
-    @patch("h2k_hpxml.utils.dependencies.installers.hpxml_installer.HPXMLInstaller.install")
-    @patch("click.echo")
-    def test_install_openstudio_hpxml_success(
-        self,
-        mock_echo,
-        mock_install,
-        manager,
-    ):
-        """Test successful OpenStudio-HPXML installation."""
-        # Mock HPXMLInstaller.install to return True (success)
-        mock_install.return_value = True
-
-        result = manager._install_openstudio_hpxml()
-
-        assert result is True
-        # Verify installer was called
-        mock_install.assert_called_once()
-
-    @patch("h2k_hpxml.utils.dependencies.installers.hpxml_installer.HPXMLInstaller.install")
-    @patch("click.echo")
-    def test_install_openstudio_hpxml_failure(self, mock_echo, mock_install, manager):
-        """Test OpenStudio-HPXML installation failure."""
-        # Mock HPXMLInstaller.install to raise an exception
-        mock_install.side_effect = Exception("Download failed")
-
-        result = manager._install_openstudio_hpxml()
-
-        assert result is False
-        mock_echo.assert_any_call("❌ OpenStudio-HPXML installation failed: Download failed")
-
-    # Removed test_update_config_file_success and test_update_config_file_not_found
-    # Config update functionality is deprecated - dependencies are now auto-detected
-
-    # Removed test_find_config_file_in_cwd and test_find_config_file_not_found
-    # Config file finding is now handled by ConfigManager, not DependencyManager
-
-    # REMOVED: test_install_openstudio_linux_debian - was causing actual OpenStudio installation damage
-    # This test was creating real DependencyManager instances that triggered file system operations
-
-    # REMOVED: test_install_openstudio_linux_tarball - was causing actual OpenStudio installation damage
-    # This test was creating real DependencyManager instances that triggered file system operations
-
-
-class TestUserInteraction:
-    """Test cases for user interaction flows."""
-
-    @pytest.fixture
-    def manager_interactive(self):
-        return DependencyManager(interactive=True, skip_deps=False)
-
-    @patch("click.prompt", return_value=3)
-    @patch("click.echo")
-    def test_handle_missing_dependencies_continue(
-        self, mock_echo, mock_prompt, manager_interactive
-    ):
-        """Test continuing without dependencies."""
-        result = manager_interactive._handle_interactive_install(False, False)
-
-        assert result is True
-        mock_echo.assert_any_call("⚠️  Continuing without all dependencies. Errors may occur.")
-
-    @patch("click.prompt", return_value=4)
-    @patch("click.echo")
-    def test_handle_missing_dependencies_exit(self, mock_echo, mock_prompt, manager_interactive):
-        """Test exiting when dependencies are missing."""
-        result = manager_interactive._handle_interactive_install(False, False)
-
-        assert result is False
-        mock_echo.assert_any_call("Installation cancelled.")
-
-    @patch("click.prompt", return_value=2)
-    @patch("h2k_hpxml.utils.dependencies.DependencyManager._show_manual_instructions")
-    def test_handle_missing_dependencies_manual(
-        self, mock_show_manual, mock_prompt, manager_interactive
-    ):
-        """Test showing manual installation instructions."""
-        result = manager_interactive._handle_interactive_install(False, False)
-
-        assert result is False
-        mock_show_manual.assert_called_once()
-
-    @patch("click.echo")
-    def test_show_manual_instructions_windows(self, mock_echo, manager_interactive):
-        """Test manual installation instructions for Windows."""
-        manager_interactive.is_windows = True
-        manager_interactive.is_linux = False
-
-        manager_interactive._show_manual_instructions(["OpenStudio", "OpenStudio-HPXML"])
-
-        echo_calls = [str(call) for call in mock_echo.call_args_list]
-        # Updated to match new portable Windows installation instructions
-        assert any("Windows.tar.gz" in call or "portable" in call for call in echo_calls)
-
-    @patch("click.echo")
-    def test_show_manual_instructions_linux(self, mock_echo, manager_interactive):
-        """Test manual installation instructions for Linux."""
-        manager_interactive.is_windows = False
-        manager_interactive.is_linux = True
-
-        manager_interactive._show_manual_instructions(["OpenStudio", "OpenStudio-HPXML"])
-
-        echo_calls = [str(call) for call in mock_echo.call_args_list]
-        assert any(".deb" in call for call in echo_calls)
-
-
-class TestUtilityFunctions:
-    """Test cases for utility functions."""
-
-    @patch("h2k_hpxml.utils.dependencies.DependencyManager.validate_all")
-    def test_validate_dependencies_function(self, mock_validate):
-        """Test the validate_dependencies convenience function."""
-        mock_validate.return_value = True
-
-        result = validate_dependencies(interactive=False, skip_deps=False, check_only=False)
-
-        assert result is True
-        mock_validate.assert_called_once()
-
-    @patch("h2k_hpxml.utils.dependencies.DependencyManager.check_only")
-    def test_validate_dependencies_check_only(self, mock_check_only):
-        """Test the validate_dependencies function with check_only=True."""
-        mock_check_only.return_value = True
-
-        result = validate_dependencies(check_only=True)
-
-        assert result is True
-        mock_check_only.assert_called_once()
-
-    @patch("h2k_hpxml.utils.dependencies.DependencyManager.validate_all")
-    def test_validate_dependencies_auto_install(self, mock_validate):
-        """Test the validate_dependencies function with install_quiet=True."""
-        mock_validate.return_value = True
-
-        result = validate_dependencies(install_quiet=True, interactive=False)
-
-        assert result is True
-        mock_validate.assert_called_once()
-
-    def test_dependency_manager_auto_install_init(self):
-        """Test DependencyManager initialization with install_quiet=True."""
-        manager = DependencyManager(install_quiet=True, interactive=False)
-
-        assert manager.install_quiet is True
-        assert manager.interactive is False
-
-    def test_validate_dependencies_with_paths(self):
-        """Test validate_dependencies function with custom paths."""
-        with patch("h2k_hpxml.utils.dependencies.cli.DependencyManager") as MockManager:
-            mock_instance = MockManager.return_value
-            mock_instance.validate_all.return_value = True
-
-            result = validate_dependencies(
-                hpxml_path="/custom/hpxml", openstudio_path="/custom/openstudio"
-            )
-
-            assert result is True
-            MockManager.assert_called_once_with(
-                interactive=True,
-                skip_deps=False,
-                install_quiet=False,
-                hpxml_path="/custom/hpxml",
-                openstudio_path="/custom/openstudio",
-            )
-
-    def test_is_debian_based_true(self):
-        """Test Debian detection when /etc/debian_version exists."""
-        from h2k_hpxml.utils.dependencies.installers.linux_installer import LinuxInstaller
-
-        installer = LinuxInstaller("3.9.0", "bb29e94a73", Path("/tmp/test"), interactive=False)
-
-        with patch("os.path.exists", return_value=True):
-            result = installer._is_debian_based()
-
-        assert result is True
-
-    def test_is_debian_based_apt_get(self):
-        """Test Debian detection when apt-get is available."""
-        from h2k_hpxml.utils.dependencies.installers.linux_installer import LinuxInstaller
-
-        installer = LinuxInstaller("3.9.0", "bb29e94a73", Path("/tmp/test"), interactive=False)
-
-        with patch("os.path.exists", return_value=False):
-            with patch("shutil.which", return_value="/usr/bin/apt-get"):
-                result = installer._is_debian_based()
-
-        assert result is True
-
-    def test_is_debian_based_false(self):
-        """Test Debian detection when neither condition is met."""
-        from h2k_hpxml.utils.dependencies.installers.linux_installer import LinuxInstaller
-
-        installer = LinuxInstaller("3.9.0", "bb29e94a73", Path("/tmp/test"), interactive=False)
-
-        with patch("os.path.exists", return_value=False):
-            with patch("shutil.which", return_value=None):
-                result = installer._is_debian_based()
-
-        assert result is False
-
-
-class TestPlatformSpecificBehavior:
-    """Test cases for platform-specific behavior."""
-
-    @pytest.mark.parametrize(
-        "platform_name,is_windows,is_linux",
-        [
-            ("Windows", True, False),
-            ("Linux", False, True),
-            ("Darwin", False, False),
-        ],
-    )
-    def test_platform_detection(self, platform_name, is_windows, is_linux):
-        """Test platform detection for different operating systems."""
-        with patch("platform.system", return_value=platform_name):
-            manager = DependencyManager()
-
-            assert manager.is_windows == is_windows
-            assert manager.is_linux == is_linux
-
-    @patch("platform.system", return_value="Windows")
-    def test_unsupported_platform_installation(self, mock_platform):
-        """Test installation failure on unsupported platform."""
-        manager = DependencyManager()
-        manager.is_windows = False
-        manager.is_linux = False
-
-        with patch("click.echo") as mock_echo:
-            result = manager._install_openstudio()
-
-        assert result is False
-        mock_echo.assert_any_call("❌ Unsupported platform: Windows")
-
-
-class TestWindowsSimulation:
-    """Test Windows-specific functionality by simulating Windows environment on Linux."""
-
-    @pytest.mark.skipif(platform.system() != "Windows", reason="Windows-specific test")
-    def test_windows_hpxml_path(self):
-        """Test Windows HPXML path generation."""
-        with patch("platform.system", return_value="Windows"):
-            manager = DependencyManager()
-            expected_path = Path("C:/OpenStudio-HPXML")
-            assert manager.default_hpxml_path == expected_path
-
-    @pytest.mark.skipif(platform.system() != "Windows", reason="Windows-specific test")
-    def test_windows_openstudio_paths(self):
-        """Test Windows OpenStudio path generation."""
-        with patch("platform.system", return_value="Windows"):
-            with patch.dict(os.environ, {"PROGRAMFILES": r"C:\Program Files"}):
-                manager = DependencyManager()
-                paths = manager._get_openstudio_paths()
-
-                # Check that Windows-specific paths are included
-                assert any("openstudio.exe" in path for path in paths)
-                assert any("Program Files" in path for path in paths)
-                assert any("C:\\openstudio\\bin\\openstudio.exe" in path for path in paths)
-
-    @pytest.mark.skipif(platform.system() != "Windows", reason="Windows-specific test")
-    def test_windows_manual_instructions(self):
-        """Test Windows manual installation instructions."""
-        with patch("platform.system", return_value="Windows"):
-            manager = DependencyManager()
-
-            with patch("click.echo") as mock_echo:
-                manager._show_openstudio_instructions()
-
-            echo_calls = [str(call) for call in mock_echo.call_args_list]
-            assert any("Windows.exe" in call for call in echo_calls)
-            assert any("administrator" in call for call in echo_calls)
-
-    @pytest.mark.skipif(platform.system() != "Windows", reason="Windows-specific test")
-    def test_windows_hpxml_instructions(self):
-        """Test Windows HPXML manual installation instructions."""
-        with patch("platform.system", return_value="Windows"):
-            manager = DependencyManager()
-
-            with patch("click.echo") as mock_echo:
-                manager._show_hpxml_instructions()
-
-            echo_calls = [str(call) for call in mock_echo.call_args_list]
-            # Should show Windows-style path separator
-            assert any("workflow\\\\run_simulation.rb" in call for call in echo_calls)
-
-    @pytest.mark.skipif(platform.system() != "Windows", reason="Windows-specific test")
-    @patch("platform.system", return_value="Windows")
-    def test_full_windows_workflow_simulation(self, mock_platform):
-        """Test complete Windows dependency workflow simulation."""
-        manager = DependencyManager(interactive=False, auto_install=False)
-
-        # Test that Windows detection works
-        assert manager.is_windows is True
-        assert manager.is_linux is False
-
-        # Test path generation
-        assert manager.default_hpxml_path == Path("C:/OpenStudio-HPXML")
-
-        # Test that installation routing works for Windows
-        with patch.object(
-            manager, "_install_openstudio_windows", return_value=False
-        ) as mock_install:
-            result = manager._install_openstudio()
-            mock_install.assert_called_once()
-            assert result is False
-
-
-class TestWindowsIntegration:
-    """Integration-style tests for Windows functionality."""
-
-    def test_windows_url_construction(self):
-        """Test Windows download URL construction."""
-        with patch("platform.system", return_value="Windows"):
-            manager = DependencyManager()
-
-            # Test URL construction matches expected pattern
-            expected_base = manager.OPENSTUDIO_BASE_URL
-            expected_hash = manager.OPENSTUDIO_BUILD_HASH
-
-            assert "github.com/NREL/OpenStudio" in expected_base
-            assert expected_hash == "c77fbb9569"
-
-            # The actual URL would be constructed in _install_openstudio_windows
-            expected_url = (
-                f"{expected_base}/"
-                f"v{manager.REQUIRED_OPENSTUDIO_VERSION}/"
-                f"OpenStudio-{manager.REQUIRED_OPENSTUDIO_VERSION}+"
-                f"{expected_hash}-Windows.exe"
-            )
-
-            assert "Windows.exe" in expected_url
-            assert manager.REQUIRED_OPENSTUDIO_VERSION in expected_url
-
-    def test_environment_simulation(self):
-        """Test simulating Windows environment variables."""
-        win_env = {
-            "PROGRAMFILES": r"C:\Program Files",
-            "PROGRAMFILES(X86)": r"C:\Program Files (x86)",
-            "OPENSTUDIO_HPXML_PATH": r"C:\MyCustom\OpenStudio-HPXML",
-        }
-
-        with patch("platform.system", return_value="Windows"):
-            with patch.dict(os.environ, win_env):
-                DependencyManager()
-
-                # Test custom HPXML path from environment
-                with patch("pathlib.Path.exists", return_value=True):
-                    with patch("pathlib.Path.__truediv__") as mock_div:
-                        mock_workflow = Mock()
-                        mock_workflow.exists.return_value = True
-                        mock_div.return_value = mock_workflow
-
-                        # This would use the environment variable path
-                        env_path = os.environ.get("OPENSTUDIO_HPXML_PATH")
-                        assert env_path == r"C:\MyCustom\OpenStudio-HPXML"
-
-
-class TestDownloadURLValidation:
-    """
-    Integration tests to verify actual download URLs are accessible.
-
-    These tests use the same SSL configuration as the production installer
-    (SSL verification bypassed) to ensure consistent behavior in corporate
-    networks and development environments with certificate interception.
-    """
-
-    @pytest.fixture
-    def manager(self):
-        return DependencyManager(interactive=False, skip_deps=False)
-
-    def _check_url_accessible(self, url, timeout=10):
-        """
-        Check if a URL is accessible without downloading the full file.
-
-        Uses the same SSL configuration as the production installer to handle
-        corporate networks with self-signed certificates or SSL interception.
-
-        Returns:
-            tuple: (is_accessible, status_info)
-        """
-        import ssl
-
-        try:
-            # Create SSL context that doesn't verify certificates (matches installer.py behavior)
-            # This is necessary for corporate networks and development environments
-            ctx = ssl.create_default_context()
-            ctx.check_hostname = False
-            ctx.verify_mode = ssl.CERT_NONE
-
-            # Use HEAD request to check if URL exists without downloading
-            request = urllib.request.Request(url, method="HEAD")
-            request.add_header("User-Agent", "h2k_hpxml-test/1.0")
-
-            # Create opener with SSL context (matches production installer)
-            opener = urllib.request.build_opener(urllib.request.HTTPSHandler(context=ctx))
-
-            with opener.open(request, timeout=timeout) as response:
-                return (
-                    True,
-                    f"Status: {response.status}, Size: {response.headers.get('content-length', 'unknown')} bytes",
-                )
-
-        except urllib.error.HTTPError as e:
-            return False, f"HTTP Error {e.code}: {e.reason}"
-        except urllib.error.URLError as e:
-            return False, f"URL Error: {e.reason}"
-        except TimeoutError:
-            return False, f"Timeout after {timeout} seconds"
-        except Exception as e:
-            return False, f"Unexpected error: {e}"
-
-    @pytest.mark.integration
-    def test_openstudio_windows_url_accessible(self, manager):
-        """Test that OpenStudio Windows installer URL is accessible."""
-        installer_url = f"{manager.OPENSTUDIO_BASE_URL}/v{manager.REQUIRED_OPENSTUDIO_VERSION}/OpenStudio-{manager.REQUIRED_OPENSTUDIO_VERSION}+c77fbb9569-Windows.exe"
-
-        accessible, info = self._check_url_accessible(installer_url)
-
-        assert (
-            accessible
-        ), f"OpenStudio Windows installer URL not accessible: {installer_url}\nError: {info}"
-        print(f"✅ Windows installer URL accessible: {info}")
-
-    @pytest.mark.integration
-    def test_openstudio_ubuntu_deb_url_accessible(self, manager):
-        """Test that OpenStudio Ubuntu .deb package URL is accessible."""
-        deb_url = f"{manager.OPENSTUDIO_BASE_URL}/v{manager.REQUIRED_OPENSTUDIO_VERSION}/OpenStudio-{manager.REQUIRED_OPENSTUDIO_VERSION}+c77fbb9569-Ubuntu-22.04-x86_64.deb"
-
-        accessible, info = self._check_url_accessible(deb_url)
-
-        assert accessible, f"OpenStudio Ubuntu .deb URL not accessible: {deb_url}\nError: {info}"
-        print(f"✅ Ubuntu .deb URL accessible: {info}")
-
-    @pytest.mark.integration
-    def test_openstudio_linux_tarball_url_accessible(self, manager):
-        """Test that OpenStudio Linux tarball URL is accessible."""
-        tarball_url = f"{manager.OPENSTUDIO_BASE_URL}/v{manager.REQUIRED_OPENSTUDIO_VERSION}/OpenStudio-{manager.REQUIRED_OPENSTUDIO_VERSION}+c77fbb9569-Ubuntu-22.04-x86_64.tar.gz"
-
-        accessible, info = self._check_url_accessible(tarball_url)
-
-        assert (
-            accessible
-        ), f"OpenStudio Linux tarball URL not accessible: {tarball_url}\nError: {info}"
-        print(f"✅ Linux tarball URL accessible: {info}")
-
-    @pytest.mark.integration
-    def test_openstudio_hpxml_url_accessible(self, manager):
-        """Test that OpenStudio-HPXML zip file URL is accessible."""
-        hpxml_url = f"{manager.HPXML_BASE_URL}/{manager.REQUIRED_HPXML_VERSION}/OpenStudio-HPXML-{manager.REQUIRED_HPXML_VERSION}.zip"
-
-        accessible, info = self._check_url_accessible(hpxml_url)
-
-        assert accessible, f"OpenStudio-HPXML URL not accessible: {hpxml_url}\nError: {info}"
-        print(f"✅ OpenStudio-HPXML URL accessible: {info}")
-
-    @pytest.mark.integration
-    def test_github_releases_pages_accessible(self, manager):
-        """Test that the GitHub releases pages are accessible."""
-        openstudio_releases = f"https://github.com/NREL/OpenStudio/releases/tag/v{manager.REQUIRED_OPENSTUDIO_VERSION}"
-        hpxml_releases = f"https://github.com/NREL/OpenStudio-HPXML/releases/tag/{manager.REQUIRED_HPXML_VERSION}"
-
-        # Test OpenStudio releases page
-        accessible, info = self._check_url_accessible(openstudio_releases)
-        assert (
-            accessible
-        ), f"OpenStudio releases page not accessible: {openstudio_releases}\nError: {info}"
-
-        # Test OpenStudio-HPXML releases page
-        accessible, info = self._check_url_accessible(hpxml_releases)
-        assert (
-            accessible
-        ), f"OpenStudio-HPXML releases page not accessible: {hpxml_releases}\nError: {info}"
-
-        print("✅ GitHub releases pages accessible")
-
-    @pytest.mark.integration
-    def test_download_url_construction(self, manager):
-        """Test that download URLs are constructed correctly based on the pattern."""
-        # Test the URL construction logic matches what the methods actually build
-        expected_windows_url = f"{manager.OPENSTUDIO_BASE_URL}/v{manager.REQUIRED_OPENSTUDIO_VERSION}/OpenStudio-{manager.REQUIRED_OPENSTUDIO_VERSION}+c77fbb9569-Windows.exe"
-        expected_hpxml_url = f"{manager.HPXML_BASE_URL}/{manager.REQUIRED_HPXML_VERSION}/OpenStudio-HPXML-{manager.REQUIRED_HPXML_VERSION}.zip"
-
-        # Verify these match what the actual methods would construct
-        assert manager.REQUIRED_OPENSTUDIO_VERSION == "3.9.0"
-        assert manager.REQUIRED_HPXML_VERSION == "v1.9.1"
-
-        # Check URL patterns
-        assert "github.com/NREL/OpenStudio/releases/download" in expected_windows_url
-        assert "github.com/NREL/OpenStudio-HPXML/releases/download" in expected_hpxml_url
-        assert manager.REQUIRED_HPXML_VERSION in expected_hpxml_url
-
-        print("✅ URL construction logic verified")
-
-    @pytest.mark.integration
-    @pytest.mark.skipif(
-        not os.getenv("NETWORK_TESTS"),
-        reason="Network tests disabled (set NETWORK_TESTS=1 to enable)",
-    )
-    def test_partial_download_verification(self, manager):
-        """Test partial downloads to verify file integrity."""
-        hpxml_url = f"{manager.HPXML_RELEASES_URL}/{manager.REQUIRED_HPXML_VERSION}/OpenStudio-HPXML-{manager.REQUIRED_HPXML_VERSION}.zip"
-
-        try:
-            # Download first 1KB to verify it's actually a zip file
-            request = urllib.request.Request(hpxml_url)
-            request.add_header("Range", "bytes=0-1023")  # First 1KB
-            request.add_header("User-Agent", "h2k_hpxml-test/1.0")
-
-            with urllib.request.urlopen(request, timeout=30) as response:
-                data = response.read()
-
-                # Check for ZIP file signature
-                assert len(data) > 0, "No data received"
-                assert (
-                    data[:4] == b"PK\x03\x04"
-                ), f"File does not appear to be a ZIP file. First 4 bytes: {data[:4]}"
-
-                print(
-                    f"✅ Partial download verified - file appears to be valid ZIP ({len(data)} bytes downloaded)"
-                )
-
-        except urllib.error.HTTPError as e:
-            if e.code == 416:  # Range not satisfiable - server doesn't support range requests
-                pytest.skip("Server doesn't support range requests, skipping partial download test")
-            else:
-                pytest.fail(f"HTTP error during partial download: {e}")
-        except Exception as e:
-            pytest.fail(f"Error during partial download verification: {e}")
-
-
-if __name__ == "__main__":
-    pytest.main([__file__, "-v"])
